v1.0.1.9 close #9 fix #11 重写抛出异常方式。 为Message建立文件夹以分类。 完善和简化一些Summary。
