+++ Mirai-CSharp.Example/Mirai-CSharp.Example.csproj
@@ -7,7 +7,7 @@
   </PropertyGroup>
   
   <ItemGroup>
-    <PackageReference Include="Mirai-CSharp" Version="1.0.1.8" />  
+    <PackageReference Include="Mirai-CSharp" Version="1.0.1.9" />  
   </ItemGroup>
   
 </Project>
