+++ Mirai-CSharp/Models/FriendApplyActions.cs
@@ -14,7 +14,7 @@
         /// </summary>
         Deny = 1,
         /// <summary>
-        /// 拒绝添加好友并添加黑名单，不再接收该用户的好友申请
+        /// 拒绝添加好友并添加黑名单, 不再接收该用户的好友申请
         /// </summary>
         DenyAndBlock = 2
     }
