+++ Mirai-CSharp/Models/GroupApplyActions.cs
@@ -18,11 +18,11 @@
         /// </summary>
         Ignore = 2,
         /// <summary>
-        /// 拒绝入群并添加黑名单，不再接收该用户的入群申请
+        /// 拒绝入群并添加黑名单, 不再接收该用户的入群申请
         /// </summary>
         DenyAndBlock = 3,
         /// <summary>
-        /// 忽略入群并添加黑名单，不再接收该用户的入群申请
+        /// 忽略入群并添加黑名单, 不再接收该用户的入群申请
         /// </summary>
         IgnoreAndBlock = 4
     }
