+++ Mirai-CSharp/Session/MiraiHttpSession.Validation.cs
@@ -11,70 +11,40 @@ namespace Mirai_CSharp
     public partial class MiraiHttpSession
     {
         /// <summary>
-        /// 通过状态码抛出相应的异常
+        /// 通过状态码返回相应的异常
         /// </summary>
-        /// <exception cref="InvalidAuthKeyException"/>
-        /// <exception cref="BotNotFoundException"/>
-        /// <exception cref="InvalidSessionException"/>
-        /// <exception cref="TargetNotFoundException"/>
-        /// <exception cref="FileNotFoundException"/>
-        /// <exception cref="PermissionDeniedException"/>
-        /// <exception cref="BotMutedException"/>
-        /// <exception cref="MessageTooLongException"/>
-        /// <exception cref="ArgumentException"/>
-        /// <exception cref="UnknownResponseException"/>
-#if !NETSTANDARD2_0
-        [DoesNotReturn]
-#endif
-        private static T ThrowCommonException<T>(int code, in JsonElement root)
+        /// <returns>
+        /// 根据给定的 <paramref name="code"/> 返回下列异常之一:
+        /// <list type="bullet">
+        /// <item><term><see cref="InvalidAuthKeyException"/></term><description><paramref name="code"/> 为 1</description></item>
+        /// <item><term><see cref="BotNotFoundException"/></term><description><paramref name="code"/> 为 2</description></item>
+        /// <item><term><see cref="InvalidSessionException"/></term><description><paramref name="code"/> 为 3 或 4</description></item>
+        /// <item><term><see cref="TargetNotFoundException"/></term><description><paramref name="code"/> 为 5</description></item>
+        /// <item><term><see cref="FileNotFoundException"/></term><description><paramref name="code"/> 为 6</description></item>
+        /// <item><term><see cref="PermissionDeniedException"/></term><description><paramref name="code"/> 为 10</description></item>
+        /// <item><term><see cref="BotMutedException"/></term><description><paramref name="code"/> 为 20</description></item>
+        /// <item><term><see cref="MessageTooLongException"/></term><description><paramref name="code"/> 为 30</description></item>
+        /// <item><term><see cref="ArgumentException"/></term><description><paramref name="code"/> 为 400</description></item>
+        /// <item><term><see cref="UnknownResponseException"/></term><description>其它情况</description></item>
+        /// </list>
+        /// </returns>
+        private static Exception GetCommonException(int code, in JsonElement root)
         {
-            switch (code)
+            return code switch
             {
-                case 1:
-                    {
-                        throw new InvalidAuthKeyException();
-                    }
-                case 2:
-                    {
-                        throw new BotNotFoundException();
-                    }
-                case 3:
-                    {
-                        throw new InvalidSessionException();
-                    }
-                case 4:
-                    {
-                        throw new InvalidSessionException();
-                    }
-                case 5:
-                    {
-                        throw new TargetNotFoundException();
-                    }
-                case 6:
-                    {
-                        throw new FileNotFoundException("指定的文件不存在。");
-                    }
-                case 10:
-                    {
-                        throw new PermissionDeniedException();
-                    }
-                case 20:
-                    {
-                        throw new BotMutedException();
-                    }
-                case 30:
-                    {
-                        throw new MessageTooLongException();
-                    }
-                case 400:
-                    {
-                        throw new ArgumentException("调用http-api失败, 参数错误, 请到 https://github.com/Executor-Cheng/Mirai-CSharp/issues 下提交issue。");
-                    }
-                default:
-                    {
-                        throw new UnknownResponseException(root.GetRawText());
-                    }
-            }
+                1 => new InvalidAuthKeyException(),
+                2 => new BotNotFoundException(),
+                // 3 or 4 => new InvalidSessionException(), // C# 9.0
+                3 => new InvalidSessionException(),
+                4 => new InvalidSessionException(),
+                5 => new TargetNotFoundException(),
+                6 => new FileNotFoundException("指定的文件不存在。"),
+                10 => new PermissionDeniedException(),
+                20 => new BotMutedException(),
+                30 => new MessageTooLongException(),
+                400 => new ArgumentException("调用http-api失败, 参数错误, 请到 https://github.com/Executor-Cheng/Mirai-CSharp/issues 下提交issue。"),
+                _ => new UnknownResponseException(root.GetRawText())
+            };
         }
 
         private void CheckDisposed()
